Add options to the GKE cluster
This allows to specify the machine type to use,
whether to use preemptible VMs or not, and most
importantly the VPC and subnet in which to create
the cluster.

On-prem connectivity is often not done in the default
VPC because on-prem connectivity often requires specific
IP ranges to be used.